Read tests moved out of DataTest
